% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_partial.R
\name{bootstrap_partial}
\alias{bootstrap_partial}
\title{Bootstrapping Partial Invariance}
\usage{
bootstrap_partial(
  saved_model,
  data,
  model,
  group,
  nboot = 1000,
  invariance_index,
  invariance_rule,
  invariance_compare,
  partial_step,
  group.equal
)
}
\arguments{
\item{saved_model}{A saved lavaan model of the
step you would like to estimate partial
invariance for.}

\item{data}{The dataframe for the estimation}

\item{model}{The original model lavaan syntax}

\item{group}{The grouping variable column as a character}

\item{nboot}{The number of bootstraps you would like
to calculate. Please note: large models with many parameters
will run slowly depending on your computer.}

\item{invariance_index}{The name of the fix index you
want to use for invariance testing. For example,
"cfi", or "rmsea" - must be a character vector and
part of the \code{fitmeasures} provided by
\code{lavaan}.}

\item{invariance_rule}{The difference between your
previous model fit index and the new partial model
index that you would accept as invariant. Must be a
numeric value.}

\item{invariance_compare}{The comparison fit index to
judge if a model is invariant. This value will be compared
to the calculated partial invariance models and the
difference will be calculated to compare to \code{invariance_rule}.}

\item{partial_step}{Which parameter you would like to
bootstrap partial invariance for. You can use "loadings",
"regressions", "intercepts", "residuals", or "thresholds".
Note that for some models this may estimate more than you
want (i.e., the syntax for factor covariances is \code{~~}
as well as residuals) but these can be excluded from the
from the final dataframe.}

\item{group.equal}{The equality constraints for this
bootstrapping - use this parameter in case you want to
estimate effect size for a specific partial step
but continue to hold all other things constrained}
}
\value{
A set of graphs and dataframes.

\item{invariance_plot}{A ggplot2 object that visualizes
the bootstrapped and random estimates difference score
between groups. These are separated by "invariant" and
"non-invariant" models.}
\item{effect_invariance_plot}{A ggplot2 object that
visualizes the effect sizes of the differences of the
invariance plot in a forest plot style. }
\item{density_plot}{A ggplot2 object that
visualizes the raw data of the boot_DF comparing
groups and invariance results.}
\item{boot_DF}{A dataframe of the bootstrapped and
randomized results.}
\item{boot_summary}{A dataframe of the summary of the
bootstrapped and random results with effect sizes. Note:
these last two dataframes can be used to recreate the
visualizations in your own style. }
\item{boot_effects}{A dataframe of the summary of the
bootstrapped and random results with effect sizes for
replication rates. }
}
\description{
A function to bootstrap a partially invariant
multigroup CFA model. This function bootstraps
a model with a focus on freeing one parameter
at a time for a specific partial invariance model.
For example, you can use this function to view all
differences in intercepts for observed variables
between two groups. This function returns the
bootstrapped values, a summary table, and two
invariance plots comparing bootstrapped data to
randomized group assignment.
}
\examples{
library(lavaan)
HS.model <- ' visual  =~ x1 + x2 + x3
textual =~ x4 + x5 + x6
speed   =~ x7 + x8 + x9 '

saved_model <- cfa(HS.model, data = HolzingerSwineford1939,
 meanstructure = TRUE,
 group = "sex",
 group.equal = c("loadings"))

# not run to save load time
# saved_boot <- bootstrapped_partial(
#  saved_model = saved_model,
#  data = HolzingerSwineford1939,
#  model = HS.model,
#  group = "sex",
#  nboot = 100,
#  invariance_index = "cfi",
#  invariance_rule = .01,
#  invariance_compare = fitmeasures(saved_model, "cfi"),
#  partial_step = c("loadings"),
#  group.equal = c("loadings"))

# saved_boot$boot_DF
# saved_boot$boot_summary
# saved_boot$invariance_plot
# saved_boot$effect_invariance_plot
}
\keyword{cfa,}
\keyword{lavaan}
\keyword{multigroup}
\keyword{sem,}
