% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_rr.R
\name{bootstrap_rr}
\alias{bootstrap_rr}
\title{Bootstrapping Measurement Invariance}
\usage{
bootstrap_rr(
  saved_configural,
  nboot = 1000,
  invariance_index,
  invariance_rule,
  group.equal
)
}
\arguments{
\item{saved_configural}{A saved \code{lavaan} model
of the "configural" model that includes the grouping
variable but no other equality constraints.}

\item{nboot}{The number of bootstraps you would like
to calculate. Please note: large models with many parameters
will run slowly depending on your computer.}

\item{invariance_index}{The name of the fix index you
want to use for invariance testing. For example,
"cfi", or "rmsea" - must be a character vector and
part of the \code{fitmeasures} provided by
\code{lavaan}.}

\item{invariance_rule}{The difference between your
previous model fit index and the new equality model
index that you would accept as invariant. Must be a
numeric value.}

\item{group.equal}{The equality constraints you would
like to impose as a vector, in order. This argument
is the same as \code{lavaan} syntax. For example,
c("loadings", "intercepts", "residuals").}
}
\value{
A dataframe of the bootstrapped results.
This set includes the proportion of invariant tests
for bootstrapping and random results and the
effect size h of this difference.
}
\description{
A function to help bootstrap expected rates of
"replication" for measurement invariance as compared
to a random assignment of groups. This function is
mostly designed for estimation for registered reports
but can also be used after a model is completed to
estimate rates of expected differences in the
future.
}
\examples{
HS.model <- ' visual  =~ x1 + x2 + x3
textual =~ x4 + x5 + x6
speed   =~ x7 + x8 + x9 '

library(lavaan)

data("HolzingerSwineford1939")

saved_model <- cfa(HS.model, data = HolzingerSwineford1939,
 meanstructure = TRUE,
 group = "sex",
 group.equal = c("loadings"))

saved_configural <- cfa(HS.model, data = HolzingerSwineford1939,
 meanstructure = TRUE,
 group = "sex")

saved_boot <- bootstrap_rr(
 saved_configural = saved_configural,
 nboot = 10, # don't do this - this is to make it run fast
 invariance_index = "cfi",
 invariance_rule = .01,
 group.equal = c("loadings", "intercepts", "residuals"))

 saved_boot

}
\keyword{cfa,}
\keyword{lavaan}
\keyword{multigroup}
\keyword{sem,}
