% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_d.R
\name{calculate_d}
\alias{calculate_d}
\title{\eqn{d_s} for Between Subjects with Pooled SD Denominator}
\usage{
calculate_d(
  m1 = NULL,
  m2 = NULL,
  sd1 = NULL,
  sd2 = NULL,
  n1 = NULL,
  n2 = NULL,
  a = 0.05
)
}
\arguments{
\item{m1}{mean group one}

\item{m2}{mean group two}

\item{sd1}{standard deviation group one}

\item{sd2}{standard deviation group two}

\item{n1}{sample size group one}

\item{n2}{sample size group two}

\item{a}{significance level}
}
\value{
Provides the effect size (Cohen's *d*) with associated
central and non-central confidence intervals,
the *t*-statistic, the confidence intervals associated with the means of
each group, as well as the standard deviations and standard errors
of the means for each group. The one-tailed confidence interval
is also included for sensitivity analyses.

\item{d}{effect size}
\item{dlow}{lower level confidence interval of d value}
\item{dhigh}{upper level confidence interval of d value}
}
\description{
This function displays d for two between subjects groups
and gives the central and non-central confidence interval
using the pooled standard deviation as the denominator. Taken
from MOTE for this package. Use MOTE for full options.
}
\details{
To calculate \eqn{d_s}, mean two is subtracted from mean one and divided
by the pooled standard deviation.
\deqn{d_s = \frac{M_1 - M_2}{S_{pooled}}}
}
\examples{
calculate_d(m1 = 4, m2 = 3,
sd1 = 1, sd2 = 1,
n1 = 100, n2 = 100,
a = .05)

}
\keyword{between-subjects,}
\keyword{deviation,}
\keyword{effect}
\keyword{independent}
\keyword{pooled}
\keyword{sd}
\keyword{size,}
\keyword{standard}
\keyword{t,}
