% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgcfa.R
\name{mgcfa}
\alias{mgcfa}
\title{Multigroup Confirmatory Factor Analysis}
\usage{
mgcfa(
  model,
  data = NULL,
  sample.cov = NULL,
  sample.mean = NULL,
  sample.nobs = NULL,
  group = NULL,
  group.equal,
  group.partial = NULL,
  conf.level = conf.level,
  ...
)
}
\arguments{
\item{model}{A character string representing the
overall model formatted in \code{lavaan} syntax}

\item{data}{A dataframe that includes a grouping
variable for multigroup analysis. Define \code{data}
and \code{group} OR \code{sample.cov},
\code{sample.mean}, and \code{sample.nobs}.}

\item{sample.cov}{A sample variance-covariance
matrix that includes row and/or column names that
match the observed variables in the model. For
multigroup analyses, you need to format this data
as a list of variance-covariance matrices with
names for each group.}

\item{sample.mean}{The mean vector for observed
variables. For multigroup analyses, you need to format
this data as a list with names for each group.
It is assumed they are in the same order as
\code{sample}.cov.}

\item{sample.nobs}{Number of observations for each
group. For multigroup analyses, you need to format
this data as a list with names for each group.
It is assumed they are in the same order as
\code{sample.cov}.}

\item{group}{A character name of the group column
included in the \code{data} argument.}

\item{group.equal}{A vector of names of the
constraints to test for multigroup analysis, in
the order that they should be tested. See \code{lavaan}
for possible options.}

\item{group.partial}{A vector of constraints to relax
for partial group models.}

\item{conf.level}{The confidence interval for model
coefficients.}

\item{...}{Other arguments to be included to configure
the \code{cfa} function from \code{lavaan}. For
example, you can include arguments for ordered
models, clustering, sampling.weights, or estimators.}
}
\value{
A list of outputs from the multigroup
procedure.

\item{model_coef}{A tidy dataframe of model
coefficients}
\item{model_fit}{A tidy dataframe of model fit indices}
\item{model.overall}{If a dataframe is provided,
the overall CFA model with all groups combined is included}
\item{model.GROUPS}{The individual group models}
\item{model.configural}{The configural model that
puts together both groups in the same model with
no constraints}
\item{model.CONSTRAINTS}{The model constraints provided
in group.equal are converted into models with those
increasing equality constraints}
}
\description{
A function to calculate multigroup confirmatory
factor analysis given a list of model constraints.
Returns all models using \code{lavaan} package
and tidy dataframes of relevant information.
}
\examples{
HS.model <- ' visual  =~ x1 + x2 + x3
textual =~ x4 + x5 + x6
speed   =~ x7 + x8 + x9 '

library(lavaan)

data("HolzingerSwineford1939")

saved_mgcfa <- mgcfa(model = HS.model,
 data = HolzingerSwineford1939,
 group = "sex",
 group.equal = c("loadings", "intercepts", "residuals"),
 meanstructure = TRUE)

}
\keyword{cfa,}
\keyword{lavaan}
\keyword{multigroup}
\keyword{sem,}
