% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partial_mi.R
\name{partial_mi}
\alias{partial_mi}
\title{Calculation of Partial Invariance by Parameter}
\usage{
partial_mi(saved_model, data, model, group, group.equal, partial_step)
}
\arguments{
\item{saved_model}{A \code{lavaan} model with
at least one level of constraints across groups.}

\item{data}{Dataframe from the original saved model.}

\item{model}{Model from the original saved model.}

\item{group}{Grouping variable column from the
original saved model. Use a character vector.}

\item{group.equal}{A vector of names of the
constraints to test for multigroup analysis, in
the order that they should be tested. See \code{lavaan}
for possible options.}

\item{partial_step}{The level of partial invariance
you would like to test. You can use "loadings",
"regressions", "intercepts", "residuals", or "thresholds".
Note that for some models this may estimate more than you
want (i.e., the syntax for factor covariances is \code{~~}
as well as residuals) but these can be excluded from the
from the final dataframe.}
}
\value{
Models with each constraint relaxed individually
and a summary table of fit indices from each model.

\item{models}{Saved \code{lavaan} models for each
parameter relaxed.}
\item{fit_table}{A dataframe of fit indices for
all models to investigate places for partial invariance.}
}
\description{
This function iterates over a constrained model
and examines potential areas for partial invariance.
The function returns a list of models and a dataframe
so that the user can decide which items may be relaxed
for partial invariance.
}
\examples{
HS.model <- ' visual  =~ x1 + x2 + x3
textual =~ x4 + x5 + x6
speed   =~ x7 + x8 + x9 '

library(lavaan)

data("HolzingerSwineford1939")

saved_mgcfa <- mgcfa(model = HS.model,
 data = HolzingerSwineford1939,
 group = "sex",
 group.equal = c("loadings", "intercepts", "residuals"),
 meanstructure = TRUE)

saved_mi <- partial_mi(saved_model =
  saved_mgcfa$invariance_models$model.loadings,
  data = HolzingerSwineford1939,
  model = HS.model,
  group = "sex",
  group.equal = "loadings",
  partial_step = "loadings")

  saved_mi$fit_table
  saved_mi$models

}
\keyword{cfa,}
\keyword{lavaan}
\keyword{multigroup}
\keyword{sem,}
