% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mi.R
\name{plot_mi}
\alias{plot_mi}
\title{Visualization for Measurement Invariance}
\usage{
plot_mi(
  data_coef,
  model_step,
  item_name,
  x_limits = c(-1, 1),
  y_limits,
  conf.level = 0.95,
  model_results,
  lv_name,
  plot_groups = NULL
)
}
\arguments{
\item{data_coef}{A tidy dataframe of the coefficients
from \code{mgcfa} function in this package.}

\item{model_step}{The model step you would like to
plot using this function. Check the "Model" column
in the coefficients to pick. Include this as a
character vector.}

\item{item_name}{The name of the item/variable
you want to plot for comparison. Check out the
"term" column to know your options. Usually
observed variables from your model.}

\item{x_limits}{The limits of the latent variable
to plot on on the latent mean graph.}

\item{y_limits}{The limits of the observed variable
that you included in \code{item_name}.}

\item{conf.level}{The confidence interval level you
would like to graph.}

\item{model_results}{The saved summary from \code{cfa}
or \code{mgcfa} of the same model name you listed
in \code{model_step}.}

\item{lv_name}{The name of the latent variable you
would like to plot. Ensure this name is the same
as what is in the \code{data_coef}.}

\item{plot_groups}{An optional variable to
denote which two groups you would like to plot. Used
when more than two groups are included in the output.}
}
\value{
A list of graphs created to visualize
measurement invariance.

\item{complete}{A ggplot2 object that includes all
visualization stacked together for publication.}
\item{intercept}{A ggplot2 object that includes the
intercept section of the overall graph.}
\item{mean}{A ggplot2 object that includes the
latent means section of the overall graph.}
\item{variance}{A ggplot2 object that includes the
variance/residuals section of the overall graph.}
}
\description{
A function to create visualizations for
measurement invariance comparison between
two groups. Please note: this function graphs
two groups at a time - use the function several
times if you have more than two groups.
}
\examples{

HS.model <- ' visual  =~ x1 + x2 + x3
textual =~ x4 + x5 + x6
speed   =~ x7 + x8 + x9 '

library(lavaan)

data("HolzingerSwineford1939")

saved_model <- cfa(HS.model, data = HolzingerSwineford1939,
 meanstructure = TRUE,
 group = "sex",
 group.equal = c("loadings"))

 saved_mgcfa <- mgcfa(model = HS.model,
  data = HolzingerSwineford1939,
  group = "sex",
  group.equal = c("loadings", "intercepts", "residuals"),
  meanstructure = TRUE)

# saved_mi_plots <- plot_mi(data_coef = saved_mgcfa$model_coef,
#  model_step = "Configural", # which model
#  item_name = "x1", # name of observed item
#  x_limits = c(-1,1), # LV limits to graph
#  y_limits = c(min(HolzingerSwineford1939$x1),
#             max(HolzingerSwineford1939$x1)), # Y min and max in data
#  conf.level = .95, # what ci do you want
#  model_results = saved_mgcfa$model_configural, # what model results do you want
#  lv_name = "visual", # which latent is the observed variable on
#  plot_groups = NULL)

}
\keyword{cfa,}
\keyword{lavaan}
\keyword{multigroup}
\keyword{sem,}
